news-reader-agent/pyproject.toml
[project]
name = "news-reader-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai[tools]>=0.152.0",
    "python-dotenv>=1.1.1",
    "openai>=1.98.0",
    "langchain-openai>=1.0.3",
    "litellm>=1.79.3",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.0",
]
